requirements.txt
numpy 
pyyaml
gradio